+++ python/cudf/cudf/_lib/filling.pyx
@@ -1,7 +1,6 @@
 # Copyright (c) 2020, NVIDIA CORPORATION.
 
 import numpy as np
-from cudf._lib.scalar import as_scalar
 
 from libcpp cimport bool
 from libcpp.memory cimport unique_ptr
@@ -114,18 +113,3 @@ def sequence(int size, Scalar init, Scalar step):
         ))
 
     return Column.from_unique_ptr(move(c_result))
-
-
-def arange(start, stop=None, step=1, dtype=None):
-
-    if stop is None:
-        stop = start
-        start = 0
-
-    if step is None:
-        step = 1
-
-    size = int(np.ceil((stop - start) / step))
-
-    return sequence(
-        size, as_scalar(start, dtype=dtype), as_scalar(step, dtype=dtype))