+++ python/cudf/cudf/core/column/string.py
@@ -4421,9 +4421,7 @@ class StringColumn(column.ColumnBase):
 
         if len(children) == 0 and size != 0:
             # all nulls-column:
-            offsets = libcudf.column.scalar_to_column(
-                0, size + 1, dtype="int32"
-            )
+            offsets = utils.full(size + 1, 0, dtype="int32")
 
             chars = cudf.core.column.as_column([], dtype="int8")
             children = (offsets, chars)