+++ python/cudf/cudf/core/index.py
@@ -10,7 +10,6 @@ import pandas as pd
 from pandas._config import get_option
 
 import cudf
-import cudf._lib as libcudf
 from cudf._lib.nvtx import annotate
 from cudf.core.abc import Serializable
 from cudf.core.column import (
@@ -1512,9 +1511,7 @@ class RangeIndex(Index):
     @cached_property
     def _values(self):
         if len(self) > 0:
-            return libcudf.filling.arange(
-                self._start, self._stop, dtype=self.dtype
-            )
+            return utils.arange(self._start, self._stop, dtype=self.dtype)
         else:
             return column.column_empty(0, masked=False, dtype=self.dtype)
 
@@ -1742,7 +1739,7 @@ class RangeIndex(Index):
 
 
 def index_from_range(start, stop=None, step=None):
-    vals = libcudf.filling.arange(start, stop, step, dtype=np.int64)
+    vals = utils.arange(start, stop, step, dtype=np.int64)
     return as_index(vals)